+++ CHANGELOG.md
@@ -13,7 +13,8 @@ CHANGELOG
 * Fixed LtiServiceServer media type handling
 * Fixed LtiServiceClient grant assertion aud claim  
 * Fixed AgsClaim with not mandatory line item container url 
-* Fixed OAuth2 token validation to support multiple audiences  
+* Fixed OAuth2 token validation to support multiple audiences 
+* Fixed proctoring start assessment message validator with resource link check 
 * Updated LtiServiceServerRequestHandlerInterface signature
 * Updated Guzzle dependency to ^6.5 || ^7.0 
 * Updated documentation